Move manifest patcher to build assembly
