+++ UnityProjects/MRTKDevTemplate/Assets/BuildAssets/AndroidManifestPatcher.cs.meta
