--- UnityProjects/MRTKDevTemplate/Assets/Scripts/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4c15ba023d835404c8478e6366c833f2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
